Create key objects
Key objects allow for stronger typing of the data that goes in and out of Redis. It provides a contract for the types of values that will be stored in a given list, set, etc. It also provides the concept of key prefixes which namespace arbitrary groups of keys and allow duplicate names between them without data loss.

Key objects do not replace the existing bare command method of interaction with Redis but rather supplement it with a more tightly controlled access system available if desired.

Because they are implemented as interfaces they are very easy to mock and provide a simple way to test business logic methods in client applications that interact with redis without actually hitting the data store.

This first version of key objects does not include SortedSet or HyperLog.